Changed most abilities to frame based instead of ms based.

+++ project-colossus/Assets/Scripts/Pugilist/PugilistAttack.cs
@@ -5,9 +5,9 @@ public class PugilistAttack : PrimaryAbility {
 
     [Range(0.0f, 3.0f)]
     public int m_controller;
-    public float m_startupTime;
-    public float m_activeTime;
-    public float m_cooldownTime;
+    public int m_startupTime;
+    public int m_activeTime;
+    public int m_cooldownTime;
     public int damage;
     public int aggroGenerated;
     public float radius;
