+++ project-colossus/Assets/Scripts/Pugilist/PugilistReflect.cs
@@ -6,9 +6,9 @@ public class PugilistReflect : MonoBehaviour, IAbility
 {
 
     public int m_controller;
-    public float m_startupTime;
-    public float m_activeTime;
-    public float m_cooldownTime;
+    public int m_startupTime;
+    public int m_activeTime;
+    public int m_cooldownTime;
 
     public int damage;              // Not sure about this
     public int aggroGenerated;
